[FEAT] 리포트 삭제 메서드 별도 메서드로 분리
